refactor(DEQ-200): Extract reminders and tasks sections to fix file_length
- Create StackEditorView+Reminders.swift for reminders section (~76 lines)
- Create StackEditorView+Tasks.swift for tasks sections (~88 lines)
- StackEditorView.swift reduced from 503 to 425 lines
- StackEditorView+EditMode.swift reduced from 518 to 431 lines
- Both files now well under SwiftLint 500 line limit
